feat: blueprinted' file structure, new-tests inc
